week08/explore_phila_data/run_sql/requirements.txt
functions_framework==3.5.0
google-cloud-bigquery==3.18.0
google-cloud-storage==2.15.0
Requests==2.31.0
